+++ order/order.routes.ts
@@ -1,8 +1,6 @@
-import {Router} from 'express'
-import {validationResult} from "express-validator"
-import {Order} from '../models/Order.js'
-import {Product} from '../models/Product.js'
-import {orderValidators, productValidators} from '../utils/validators.js'
+import express, {Router} from 'express'
+import {Order} from './order.schema'
+import {Product} from '../product/product.schema'
 
 
 export const router = Router()
@@ -29,41 +27,36 @@ router.get('/:id', async (req, res) => {
     }
 })
 
-router.post('/', [orderValidators],
-    async (req, res) => {
+router.post('/',
+    async (req: express.Request, res: express.Response) => {
         try {
-            const errors = validationResult(req)
-            if (!errors.isEmpty()) {
-                return res.status(400).json({errors: errors.array()})
-            }
-
             const {user, product, qty} = req.body
             const productInstance = await Product.findById(product)
-            const sum = productInstance.price * qty
+            if (productInstance?.price) {
+                const sum = productInstance.price * qty
 
-            const order = new Order({user, product, qty, sum})
-            await order.save()
+                const order = new Order({user, product, qty, sum})
+                await order.save()
 
-            res.status(201).json('Order added')
+                res.status(201).json('Order added')
+            } else {
+                res.status(404).json({message: 'Product does not exist'})
+            }
         } catch (e) {
             res.status(404).json('Bad request')
         }
     })
 
-router.patch('/:id', [orderValidators],
-    async (req, res) => {
+router.patch('/:id',
+    async (req: express.Request, res: express.Response) => {
         try {
-            const errors = validationResult(req)
             const params = req.body
-            if (!errors.isEmpty()) {
-                for (let error of errors.array()) {
-                    if (error.param in params) {
-                        return res.status(400).json({errors: error})
-                    }
-                }
-            }
             const order = await Order.findById(req.params.id)
 
+            if (!order) {
+                res.status(404).json({message: 'Order does not exist'})
+                return
+            }
             if (params.product || params.qty) {
                 let productInstance, qty
                 if (params.product) {
@@ -75,6 +68,10 @@ router.patch('/:id', [orderValidators],
                     qty = params.qty
                 } else {
                     qty = order.qty
+                }
+                if (!productInstance || ! productInstance.price) {
+                    res.status(404).json({message: 'Product in order does not exist'})
+                    return
                 }
                 params.sum = productInstance.price * qty
             }
@@ -91,9 +88,13 @@ router.patch('/:id', [orderValidators],
 router.delete('/:id', async (req, res) => {
     try {
         const order = await Order.findById(req.params.id)
+        if (!order) {
+            res.status(404).json({message: 'Order does not exist'})
+            return
+        }
         await order.deleteOne()
 
-        res.status(200).json('Product deleted')
+        res.status(200).json('Order deleted')
     } catch (e) {
         res.status(404).json({message: 'Bad request'})
     }
