+++ product/product.routes.ts
@@ -1,8 +1,6 @@
-import {Router} from 'express'
-import {validationResult} from 'express-validator'
+import express, {Router} from 'express'
 
-import {Product} from '../models/Product.js'
-import {productValidators} from "../utils/validators.js"
+import {Product} from './product.schema'
 
 export const router = Router()
 
@@ -28,12 +26,12 @@ router.get('/:id', async (req, res) => {
     }
 })
 
-router.post('/', [productValidators],
-    async (req, res) => {
+router.post('/',
+    async (req: express.Request, res: express.Response) => {
         try {
-            const errors = validationResult(req);
-            if (!errors.isEmpty()) {
-                return res.status(400).json({errors: errors.array()});
+            const existingProduct = await Product.findOne({name: req.body.name})
+            if (existingProduct) {
+                return res.status(400).json({message: 'Such product exists'})
             }
 
             const product = new Product(req.body)
@@ -46,23 +44,18 @@ router.post('/', [productValidators],
         }
     })
 
-router.patch('/:id', [productValidators],
-    async (req, res) => {
+router.patch('/:id',
+    async (req: express.Request, res: express.Response) => {
         try {
-            const errors = validationResult(req)
             const params = req.body
-            if (!errors.isEmpty()) {
-                for (let error of errors.array()) {
-                    if (error.param in params) {
-                        return res.status(400).json({errors: error})
-                    }
-                }
-            }
             const product = await Product.findById(req.params.id)
 
+            if (!product) {
+                res.status(404).json({message: 'Product does not exist'})
+                return
+            }
             Object.assign(product, params)
             await product.save()
-
             res.status(200).json('Product updated')
         } catch (e) {
             res.status(404).json({message: 'Bad request'})
@@ -72,8 +65,11 @@ router.patch('/:id', [productValidators],
 router.delete('/:id', async (req, res) => {
     try {
         const product = await Product.findById(req.params.id)
+        if (!product) {
+            res.status(404).json({message: 'Product does not exist'})
+            return
+        }
         await product.deleteOne()
-
         res.status(200).json('Product deleted')
     } catch (e) {
         res.status(404).json({message: 'Bad request'})
